+++ packages/mwp-api-proxy-plugin/src/proxy.test.js
@@ -16,32 +16,24 @@ jest.mock('mwp-config', () => {
 });
 
 const MOCK_HAPI_REQUEST = {
-	auth: { credentials: { memberCookie: 'foo member', csrfToken: 'bar token' } },
+	auth: {
+		credentials: { memberCookie: 'foo member', csrfToken: 'bar token' },
+	},
 	headers: {},
 	query: {},
-	state: { oauth_token: 'asdfasd' },
+	method: 'get',
+	state: {
+		oauth_token: 'foo',
+	},
 	server: getServer(),
+	log: () => {},
+	trackActivity: () => {},
 	getLanguage: () => 'en-US',
 };
 
 describe('apiProxy$', () => {
 	const queries = [mockQuery(MOCK_RENDERPROPS), mockQuery(MOCK_RENDERPROPS)];
 	it('returns an observable that emits an array of results', () => {
-		const getRequest = {
-			auth: {
-				credentials: { memberCookie: 'foo member', csrfToken: 'bar token' },
-			},
-			headers: {},
-			query: {},
-			method: 'get',
-			state: {
-				oauth_token: 'foo',
-			},
-			server: getServer(),
-			log: () => {},
-			trackActivity: () => {},
-			getLanguage: () => 'en-US',
-		};
 		const requestResult = {
 			type: 'fake',
 			value: { foo: 'bar' },
@@ -52,7 +44,7 @@ describe('apiProxy$', () => {
 			requestResult
 		);
 		const expectedResults = [requestResult, requestResult];
-		return apiProxy$(getRequest)(queries)
+		return apiProxy$(MOCK_HAPI_REQUEST)(queries)
 			.toPromise()
 			.then(results => expect(results).toEqual(expectedResults));
 	});