+++ packages/mwp-tracking-plugin/src/util/avro.test.js
@@ -85,14 +85,10 @@ describe('Activity tracking', () => {
 		// create a new buffer from that string
 		const avroBuffer = new Buffer(valObj.record, 'base64');
 		// get the avro-encoded record
-		const recordedInfo = avsc.parse(avro.schemas.activity).fromBuffer(avroBuffer);
-		const expectedTrackedInfo = {
-			...trackInfo,
-			aggregratedUrl: '', // misspelled, unused field in v3 spec, default ''
-			browserId: '',
-		};
-		delete expectedTrackedInfo.sessionId; // not part of v3 spec
-		expect(recordedInfo).toEqual(expectedTrackedInfo);
+		const recordedInfo = avsc
+			.parse(avro.schemas.activity)
+			.fromBuffer(avroBuffer);
+		expect(recordedInfo).toMatchSnapshot();
 	});
 });